Circuit configuration
Contains MPU6050, ESP32, PCA9685, 3 MG996R motors, a 2-battery 18650 Battery with case, and a 5V 5A step-down converter. The MPU and PCA are connected to the ESP through 2 separate I2C busses since the MPU cannot be soldered directly into the board, since it's on the outside. Other than that, everything is soldered inside the casing. Fun project, would revisit and reiterate.